feat: refactor subscription management and add car-related GraphQL operations
- Rename `Subscription` table to `UserSubscription` and update foreign key constraints
- Introduce GraphQL queries and mutations for managing cars: GetUserCarsQuery, CreateCarMutation, RemoveCarMutation, UpdateCarMutation
- Clean up Main component by removing inline GraphQL queries and utilizing imported operations
- Update FormField component for consistent styling
